Add extract-genes script to extract features from BED file with augmented GFF features Add man pages for feature-view and extract-genes

+++ Small-test/small-test.sh
@@ -13,8 +13,15 @@ pause()
     read junk
 }
 
+if [ $0 != ./small-test.sh ]; then
+    printf "Must be run as ./small-test.sh"
+    exit
+fi
+
 rm -f small-test-augmented*.bed small-test-overlaps.tsv
+cd ..
 make clean all
+cd Small-test
 
 printf "Viewing pruned GFF input...\n"
 pause
@@ -23,7 +30,7 @@ awk '{ if ( $0 ~ "^#" ) print $0; else printf("%s\t%u\t%u\t%u\t%s(%s)\n",
 
 printf "Running peak-classifier...\n"
 pause
-./peak-classifier small-test.bed small-test.gff3 small-test-overlaps.tsv
+../peak-classifier small-test.bed small-test.gff3 small-test-overlaps.tsv
 
 printf "Viewing augmented GFF data...\n"
 pause
@@ -39,5 +46,5 @@ more small-test-overlaps.tsv
 
 printf "Filtering overlaps...\n"
 pause
-./filter-overlaps small-test-overlaps.tsv small-test-filtered.tsv intron exon
+../filter-overlaps small-test-overlaps.tsv small-test-filtered.tsv intron exon
 more small-test-filtered.tsv
